% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_numeric_2_VCF.R
\name{numeric_2_VCF}
\alias{numeric_2_VCF}
\title{Convert numeric genotypes into BEAGLE compatible VCF format}
\usage{
numeric_2_VCF(genotypes, marker_map)
}
\arguments{
\item{genotypes}{A numeric matrix or dataframe coded as c(0,1,2). Columns = markers, rows = samples/taxa}

\item{marker_map}{A dataframe with the same order as the genotypes. Column names = c("rs", chr", "position")}
}
\value{
A dataframe in VCF format with marker metadata in the first 9 columns followed by genotype calls
}
\description{
Convert numeric genotypes into BEAGLE compatible VCF format
}
