% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_numeric_2_VCF.R
\name{read_vcf}
\alias{read_vcf}
\title{This function will write a VCF to a text file in the proper format}
\usage{
read_vcf(infile, skip = 8)
}
\arguments{
\item{infile}{The name of the .vcf file to import}

\item{skip}{The number of metadata lines in the .vcf}
}
\value{
A VCF dataframe
}
\description{
This function will write a VCF to a text file in the proper format
}
