% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_numeric_2_VCF.R
\name{write_vcf}
\alias{write_vcf}
\title{This function will write a VCF to a text file in the proper format}
\usage{
write_vcf(vcf_genotypes, outfile)
}
\arguments{
\item{vcf_genotypes}{A dataframe in VCF format}

\item{outfile}{The name of the output .vcf file}
}
\value{
0 if file was created, 1 if file was not created
}
\description{
This function will write a VCF to a text file in the proper format
}
